+++ source/pcraster_old_calc/calc_maskcompressortest.cc
@@ -105,7 +105,7 @@ BOOST_AUTO_TEST_CASE(script)
   REAL4 count = 1;
   for (size_t r = 0; r < rs.nrRows(); r++) {
     for (size_t c = 0; c < rs.nrCols(); c++) {
-      mask[r][c] = (r < c) ? MV_UINT1 : (r < rs.nrCols());
+      mask[r][c] = (r < c) ? MV_UINT1 : static_cast<int>(r < rs.nrCols());
       pcr::setMV(result[r][c]);
       if (mask[r][c] == 1) {
         result[r][c] = count++;
@@ -165,7 +165,7 @@ BOOST_AUTO_TEST_CASE(_0_option)
 
   for (size_t r = 0; r < rs.nrRows(); r++) {
     for (size_t c = 0; c < rs.nrCols(); c++) {
-      mask[r][c] = (r < c) ? MV_UINT1 : (r < rs.nrCols());
+      mask[r][c] = (r < c) ? MV_UINT1 : static_cast<int>(r < rs.nrCols());
       pcr::setMV(result[r][c]);
       if (mask[r][c] == 1) {
         result[r][c] = 1;
